Analyze all .coffee files when include_paths is ./
When the `fetch_files` method is passed `./` it attempts to glob
`.//**/*.coffee` which has an extra `/` and prepends all results with
`./`.

Now instead of using `fetch_files` for `./` we grab all `.coffee` files
in the current directory.

Another change that was made is that we now `chdir` to `/code` in the
`Coffeelint` initializer.