Return No Results if there are no Analyzable files.
When running this engine from the codeclimate cli to analyze a single
file, if the file is not analyzable the coffeelint engine throws the
following error:

> Running coffeelint: Done!
> error: (CC::Analyzer::Engine::EngineFailure) engine coffeelint failed
> with status 1

The current fix prevents the engine of parsing non existing results when
running against a file that is not analyzable.

Added a spec in coffeelint_spec to validate this.